core/settings.py: Use a human-redable configuration file
This commit allows users specify their settings in a much more readable
way. Before this change, they needed to be specified in a Python
dictionary. Not that it's hard to modify anyway, but the less python code
intermingled with the actual configuration, the better.

Moreover, because I've started designing the test infrastructure in this
branch, which is not still finished (it lacks integration and
network-dependant unit tests), I'm going to find much easier to write
them having this way of specifying settings.