tests: Add first working test and a test suite runner
The test added in this commit is meant to serve as an example. Please,
do not think it covers the test cases of `crypto.aes_cbc_encrypt()`.

This is a very basic test, it has no dependencies and is very straight
forward. In order to make this suite useful, we'll need some extra
complexity that allows us write tests that are as end-to-end as possible.