tests/run_test_suite.sh: Add working infrastructure for end-to-end tests
The only existing test we've got at the moment is a unit test that can be
run stand-alone as it doesn't depend on any external source. However,
because this branch is going to change some code so that it uses a
configuration file in two different ways, we'll need to start using an
end-to-end test to prove that our changes actually work.

Now, these tests will be comprised by bash script files with a very
specific structure:

1) Every file that contains end-to-end tests must have a name in the form
   "test-*.sh"

2) The tests in these files will be defined as functions with names in the
   form "test_*"

Therefore, if helper functions that don't meet those requirements are
declared, they won't be considered as tests.